address remaining comments. remove parser file
