+++ src/util/bson/indexed_map.hpp
@@ -37,10 +37,10 @@ public:
     class iterator
     {
         size_t m_idx = 0;
-        IndexedMap* m_map;
+        const IndexedMap* m_map;
     public:
 
-        iterator(IndexedMap* map, size_t idx);
+        iterator(const IndexedMap* map, size_t idx);
         entry operator*()
         {
             return m_map->operator[](m_idx);
@@ -63,15 +63,17 @@ public:
     size_t size() const;
 
     /// Find an entry by index
-    entry operator[](size_t idx);
+    entry operator[](size_t idx) const;
 
-    iterator begin();
+    iterator begin() const;
 
-    iterator end();
+    iterator end() const;
 
-    iterator front();
+    entry front();
 
-    iterator back();
+    entry back();
+
+    iterator find(const std::string& k);
 
     /// Find or add a given key
     T& operator[](const std::string& k);
@@ -85,41 +87,41 @@ public:
     const std::vector<std::string> keys() const;
 private:
     template <typename V>
-    friend inline bool operator==(const IndexedMap<V>& lhs, const IndexedMap<V>& rhs) noexcept;
+    friend bool operator==(const IndexedMap<V>& lhs, const IndexedMap<V>& rhs) noexcept;
     template <typename V>
-    friend inline bool operator!=(const IndexedMap<V>& lhs, const IndexedMap<V>& rhs) noexcept;
+    friend bool operator!=(const IndexedMap<V>& lhs, const IndexedMap<V>& rhs) noexcept;
     std::unordered_map<std::string, T> m_map;
     std::vector<std::string> m_keys;
 };
 
 template <typename T>
-inline bool operator==(const typename IndexedMap<T>::iterator& lhs,
-                       const typename IndexedMap<T>::iterator& rhs) noexcept
+bool operator==(const typename IndexedMap<T>::iterator& lhs,
+                const typename IndexedMap<T>::iterator& rhs) noexcept
 {
     return lhs.m_idx == rhs.m_idx;
 }
 
 template <typename T>
-inline bool operator!=(const typename IndexedMap<T>::iterator lhs,
-                       const typename IndexedMap<T>::iterator rhs) noexcept
+bool operator!=(const typename IndexedMap<T>::iterator lhs,
+                const typename IndexedMap<T>::iterator rhs) noexcept
 {
     return !(lhs.m_idx == rhs.m_idx);
 }
 
 template <typename T>
-inline bool operator==(const IndexedMap<T>& lhs, const IndexedMap<T>& rhs) noexcept
+bool operator==(const IndexedMap<T>& lhs, const IndexedMap<T>& rhs) noexcept
 {
     return lhs.m_map == rhs.m_map && lhs.m_keys == rhs.m_keys;
 }
 
 template <typename T>
-inline bool operator!=(const IndexedMap<T>& lhs, const IndexedMap<T>& rhs) noexcept
+bool operator!=(const IndexedMap<T>& lhs, const IndexedMap<T>& rhs) noexcept
 {
     return lhs.m_map != rhs.m_map && lhs.m_keys != rhs.m_keys;
 }
 
 template <typename T>
-IndexedMap<T>::iterator::iterator(IndexedMap<T>* map, size_t idx) : m_idx(idx), m_map(map) {}
+IndexedMap<T>::iterator::iterator(const IndexedMap<T>* map, size_t idx) : m_idx(idx), m_map(map) {}
 
 template <typename T>
 typename IndexedMap<T>::iterator& IndexedMap<T>::iterator::operator++()
@@ -138,69 +140,75 @@ typename IndexedMap<T>::iterator& IndexedMap<T>::iterator::operator--()
 template <typename T>
 typename IndexedMap<T>::iterator IndexedMap<T>::iterator::operator++(int)
 {
-    return IndexedMap<T>::iterator(m_idx++);
+    return IndexedMap<T>::iterator(m_map, m_idx++);
 }
 
 template <typename T>
 typename IndexedMap<T>::iterator IndexedMap<T>::iterator::operator--(int)
 {
-    return --(*this);
+    return IndexedMap<T>::iterator(m_map, m_idx--);
 }
-
-template <typename T>
-using entry = std::pair<std::string, T>;
 
 template <typename T>
 IndexedMap<T>::IndexedMap() {}
 
 template <typename T>
 IndexedMap<T>::IndexedMap(std::initializer_list<entry> entries) {
-    for (auto entry : entries) {
+    for (auto& entry : entries) {
         m_keys.push_back(entry.first);
         m_map[entry.first] = entry.second;
     }
 }
 
-/// The size of the map
 template <typename T>
 size_t IndexedMap<T>::IndexedMap::size() const
 {
     return m_map.size();
 }
 
-/// Find an entry by index
 template <typename T>
-entry<T> IndexedMap<T>::operator[](size_t idx)
+std::pair<std::string, T> IndexedMap<T>::operator[](size_t idx) const
 {
     auto key = m_keys[idx];
-    return { key, m_map[key] };
+    return { key, m_map.at(key) };
 }
 
 template <typename T>
-typename IndexedMap<T>::iterator IndexedMap<T>::begin()
+typename IndexedMap<T>::iterator IndexedMap<T>::begin() const
 {
     return iterator(this, 0);
 }
 
 template <typename T>
-typename IndexedMap<T>::iterator IndexedMap<T>::end()
+typename IndexedMap<T>::iterator IndexedMap<T>::end() const
 {
     return iterator(this, m_map.size());
 }
 
 template <typename T>
-typename IndexedMap<T>::iterator IndexedMap<T>::front()
+std::pair<std::string, T> IndexedMap<T>::front()
 {
-    return this->begin();
+    return *this->begin();
 }
 
 template <typename T>
-typename IndexedMap<T>::iterator IndexedMap<T>::back()
+std::pair<std::string, T> IndexedMap<T>::back()
 {
-    return this->end()--;
+    return *(--this->end());
 }
 
-/// Find or add a given key
+template <typename T>
+typename IndexedMap<T>::iterator IndexedMap<T>::find(const std::string& k)
+{
+    auto it = begin();
+    while (it != end()) {
+        if ((*it).first == k)
+            return it;
+        it++;
+    }
+    return it;
+}
+
 template <typename T>
 T& IndexedMap<T>::operator[](const std::string& k)
 {
@@ -212,14 +220,12 @@ T& IndexedMap<T>::operator[](const std::string& k)
     return m_map[k];
 }
 
-/// Whether or not this map is empty
 template <typename T>
 bool IndexedMap<T>::empty()
 {
     return m_map.empty();
 }
 
-/// Pop the last entry of the map
 template <typename T>
 void IndexedMap<T>::pop_back()
 {
