+++ tests/bson.cpp
@@ -1,12 +1,30 @@
-#include "catch2/catch.hpp"
+/*************************************************************************
+*
+* Copyright 2020 Realm Inc.
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+* http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*
+**************************************************************************/
+
 #include <stdio.h>
 #include <json.hpp>
+#include <variant>
+#include <any>
+
+#include "catch2/catch.hpp"
 #include "util/test_utils.hpp"
 #include "util/test_file.hpp"
 #include "util/bson/bson.hpp"
-#include "util/bson/parser.hpp"
-#include <variant>
-#include <any>
 
 using namespace nlohmann;
 using namespace realm;
@@ -386,53 +404,4 @@ TEST_CASE("canonical_extjson_corpus", "[bson]") {
     }
 
     // TODO: Rest of corpus
-}
-
-TEST_CASE("extjson_parser", "[bson]") {
-    struct TestParser : bson::detail::Parser {
-        using instruction = State;
-
-        std::stack<Instruction>& instructions()
-        {
-            return m_instructions;
-        }
-
-        std::stack<BsonContainer>& marks()
-        {
-            return m_marks;
-        }
-    };
-
-    auto parser = TestParser();
-
-    SECTION("null") {
-        parser.marks().push({});
-        parser.instructions().push({TestParser::instruction::JsonKey, "key"});
-        parser.null();
-        CHECK(parser.instructions().empty());
-        CHECK(parser.marks().top().back().first == "key");
-        CHECK(std::get<Null>(parser.marks().top().back().second) == bson::null);
-    }
-
-    SECTION("boolean") {
-        for (auto& b : {true, false}) {
-            parser.marks().push({});
-            parser.instructions().push({TestParser::instruction::JsonKey, "key"});
-            parser.boolean(b);
-            CHECK(parser.instructions().empty());
-            CHECK(parser.marks().top().back().first == "key");
-            CHECK(std::get<bool>(parser.marks().top().back().second) == b);
-        }
-    }
-
-    SECTION("number_integer") {
-        parser.instructions().push({TestParser::instruction::NumberInt, "key"});
-        CHECK_THROWS(parser.number_integer(42));
-    }
-
-    SECTION("number_unsigned") {
-        parser.instructions().push({});
-    }
-
-    // TODO: rest of parser methods
 }
